Change type to float

+++ function/area_circle.py
@@ -3,5 +3,5 @@ def areacircle(r):
     result = pi * r**2
     return result
 
-Radius = int(input('Enter Radius'))
+Radius = float(input('Enter Radius'))
 print('The area of the circle is ',areacircle(Radius))
